+++ src/framework/piral-core/src/components/Mediator.tsx
@@ -17,20 +17,26 @@ export interface MediatorProps {
  * The Mediator component for interfacing with pilets loading.
  */
 export const Mediator: React.FC<MediatorProps> = ({ options }) => {
-  const { initialize, readState } = useGlobalStateContext();
+  const { initialize, readState, emit } = useGlobalStateContext();
 
   React.useEffect(() => {
-    const shouldLoad = readState(s => s.app.loading);
+    const shouldLoad = readState((s) => s.app.loading);
 
     if (shouldLoad) {
       const { connect, disconnect } = startLoadingPilets(options);
+
+      emit('loading-pilets', { options });
+
       const notifier: PiletsLoading = (error, pilets, loaded) => {
         initialize(!loaded, error, pilets);
+
+        if (loaded) {
+          emit('loaded-pilets', { pilets, error });
+        }
       };
       connect(notifier);
       return () => disconnect(notifier);
     }
-
   }, none);
 
   // tslint:disable-next-line:no-null-keyword